View Source.
I know HTML includes JavaScript.  The UI is simple, and people love it.  The web API that pulls the data from the dispatch application is complex.  The public is not allowed to be visible to the internals of the TriTech System.  I'm working with Vue and Angular.  I am still unsure how to hook up a JavaScript UI tool with the MVC architecture of those systems.